Kvik now retrieves the pathway as JSON
Previously it used a websocket connection to retrieve the nodes. Now we
just send a JSON with the nodes and edges and that.

The previous implementation had the Kvik browser connect at a random
port to retrieve a pathway graph, but since we must know about every
port that has to be open to the docker container this doesn't work
anymore. Now every pathway is sent from the frontend at a specific port,
making it possible to put it into a docker container.