add new r interface prototype
The idea behind this one is that we have a pool of
n workers each with its own R instance running. The old one had
to start up a new session for every request, about 200ms
of overhead for each request. Hopefully this one will prove better.
Not fully implemented yet, but on its way.